+++ tests/test_app/forms.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from django import forms
 
 from djangocms_text_ckeditor.fields import HTMLFormField