+++ tests/test_app/models.py
@@ -1,11 +1,6 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from django.db import models
 
 from cms.models import CMSPlugin
-
-from six import python_2_unicode_compatible
 
 from djangocms_text_ckeditor.fields import HTMLField
 
@@ -14,7 +9,6 @@ class SimpleText(models.Model):
     text = HTMLField(blank=True)
 
 
-@python_2_unicode_compatible
 class DummyLink(CMSPlugin):
     label = models.TextField()
 
@@ -25,7 +19,6 @@ class DummyLink(CMSPlugin):
         return 'dummy link object'
 
 
-@python_2_unicode_compatible
 class DummySpacer(CMSPlugin):
     class Meta:
         abstract = False