+++ tests/test_plugin.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import copy
 import json
 import re
@@ -15,6 +14,8 @@ from django.utils.http import urlencode, urlunquote
 from cms.api import add_plugin, create_page, create_title
 from cms.models import CMSPlugin, Page, Title
 from cms.utils.urlutils import admin_reverse
+
+from tests.test_app.cms_plugins import DummyChildPlugin, DummyParentPlugin
 
 from djangocms_text_ckeditor.cms_plugins import TextPlugin
 from djangocms_text_ckeditor.compat import get_page_placeholders
@@ -635,6 +636,36 @@ class PluginActionsTestCase(BaseTestCase):
             self.assertEqual(response.status_code, 400)
             self.assertEqual(force_text(response.content), 'Unable to process your request.')
 
+    def test_custom_ckeditor_body_css_classes(self):
+        simple_page = create_page('test page', 'page.html', u'en')
+        simple_placeholder = get_page_placeholders(simple_page, 'en').get(slot='content')
+
+        parent_plugin = add_plugin(
+            simple_placeholder,
+            DummyParentPlugin,
+            'en',
+            label=DummyParentPlugin._ckeditor_body_class_label_trigger,
+        )
+        child_plugin = add_plugin(
+            simple_placeholder,
+            DummyChildPlugin,
+            'en',
+            target=parent_plugin,
+        )
+        text_plugin = add_plugin(
+            simple_placeholder,
+            'TextPlugin',
+            'en',
+            body="Content",
+            target=child_plugin,
+        )
+
+        with self.login_user_context(self.get_superuser()):
+            change_endpoint = self.get_change_plugin_uri(text_plugin)
+            response = self.client.get(change_endpoint)
+            self.assertContains(response, DummyParentPlugin._ckeditor_body_class)
+            self.assertContains(response, DummyChildPlugin.child_ckeditor_body_css_class)
+
     def test_render_plugin(self):
         simple_page = create_page('test page', 'page.html', u'en')
         simple_placeholder = get_page_placeholders(simple_page, 'en').get(slot='content')
@@ -813,7 +844,7 @@ class PluginActionsTestCase(BaseTestCase):
 )
 class DjangoCMSTranslationsIntegrationTestCase(BaseTestCase):
     def setUp(self):
-        super(DjangoCMSTranslationsIntegrationTestCase, self).setUp()
+        super().setUp()
         self.page = create_page('test page', 'page.html', 'en', published=True)
         self.placeholder = get_page_placeholders(self.page, 'en').get(slot='content')